requirements.txt
amqp==5.1.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dependency-injector==4.48.1
dnspython==2.6.1
email-validator==2.0.0.post2
exceptiongroup==1.2.2
fastapi==0.112.0
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
kombu==5.4.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.11
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
redis==5.0.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.15
sqlmodel==0.0.22
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.22.0
vine==5.1.0
wcwidth==0.2.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
